+++ lib/spf/context.ex
@@ -145,35 +145,40 @@ defmodule Spf.Context do
 
   @spec ipt_update({prefix, iptval}, t) :: t
   defp ipt_update({k, v}, ctx) do
+    q = elem(v, 0)
+    notq = fn {qq, _, _} -> qq != q end
     # less specific entries (if any)
     less = Iptrie.less(ctx.ipt, k) |> ipt_values(k)
     less_n = length(less)
     less_t = Enum.map(less, &elem(&1, 2)) |> Enum.join(", ")
     less_q = Enum.map([v | less], &elem(&1, 0)) |> MapSet.new() |> MapSet.size()
+    less_i = Enum.filter(less, notq) |> Enum.map(&elem(&1, 2)) |> Enum.join(", ")
 
     # more specific entries (if any)
     more = Iptrie.more(ctx.ipt, k) |> ipt_values(k)
     more_n = length(more)
     more_t = Enum.map(more, &elem(&1, 2)) |> Enum.join(", ")
     more_q = Enum.map([v | more], &elem(&1, 0)) |> MapSet.new() |> MapSet.size()
+    more_i = Enum.filter(more, notq) |> Enum.map(&elem(&1, 2)) |> Enum.join(", ")
 
     # same prefix entries (if any) -> [{q, nth, "term"}]
     other = Iptrie.get(ctx.ipt, k, {k, []}) |> elem(1)
     other_n = length(other)
     other_t = Enum.map(other, &elem(&1, 2)) |> Enum.reverse() |> Enum.join(", ")
     other_q = Enum.map([v | other], &elem(&1, 0)) |> MapSet.new() |> MapSet.size()
+    other_i = Enum.filter(other, notq) |> Enum.map(&elem(&1, 2))
 
     t = elem(v, 2)
 
     ctx
     |> Map.put(:ipt, Iptrie.put(ctx.ipt, k, [v | other]))
     |> log(:ipt, :debug, "#{t} - adds #{k} -> #{inspect(v)}")
-    |> test(:ipt, :warn, other_n > 0, "#{t} - multiple entries, already have #{k} -> #{other_t}")
-    |> test(:ipt, :warn, other_q > 1, "#{t} - inconsistent with #{other_t}")
+    |> test(:ipt, :warn, other_n > 0, "#{t} - redundant entry, already have: #{other_t}")
+    |> test(:ipt, :warn, other_q > 1, "#{t} - inconsistent with #{other_i}")
     |> test(:ipt, :warn, less_n > 0, "#{t} - unreachable due to less specific #{less_t}")
-    |> test(:ipt, :warn, less_q > 1, "#{t} - inconsistent with less specific #{less_t}")
+    |> test(:ipt, :warn, less_q > 1, "#{t} - inconsistent with less specific #{less_i}")
     |> test(:ipt, :warn, more_n > 0, "#{t} - overlaps with more specific #{more_t}")
-    |> test(:ipt, :warn, more_q > 1, "#{t} - inconsistent with more specific #{more_t}")
+    |> test(:ipt, :warn, more_q > 1, "#{t} - inconsistent with more specific #{more_i}")
   end
 
   @spec opt_ip(t, Keyword.t()) :: t