+++ CHANGELOG.md
@@ -8,11 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
-- [ ] add repeated-term warning for repeated term in same SPF
-- [ ] add repeated-term warning across SPF's as well
+- [x] change multiple entries warning to redundant-warning
 - [ ] add flag to just create a dot-file that visualizes the domain's SPF (?)
 
-- [ ] change inconsistent warning to report only the terms inconsistent with current term
+- [x] change inconsistent warning to report only the terms inconsistent with current term
 - [ ] maybe change warnings based on more/less specifics so reported source terms are unique
       At the moment an a-term with N-ip's is reporten N-times once its /24 is
       also a term.